Mirsa Rule 10.3 violations (#492)
* MISRA: fix Rule 10.3 vilations

* MISRA: Fix more Rule 10.3 vilations

* Style: fix format errors

* Spell: fix variable name

* Fix PR comments

* Fix: variable names

* Style: fix format

* Fix PR comments

* Fix variable check in cache

* Remove non needed check

* Coverage: fix coverage

* test commit

* Test commit

* Remove print statements

* Initialize some memeory

* Testing workflow

* Testing on workflow

* Fix: ut coverage

* Revert workflow changes

* Style: fix format error

+++ source/FreeRTOS_IP_Utils.c
@@ -126,7 +126,7 @@ static NetworkBufferDescriptor_t * prvPacketBuffer_to_NetworkBuffer( const void 
     {
         IPStackEvent_t xEventMessage;
         const TickType_t uxDontBlock = 0U;
-        uintptr_t uxOption = eGetDHCPState();
+        uintptr_t uxOption = ( uintptr_t ) eGetDHCPState();
 
         xEventMessage.eEventType = eDHCPEvent;
         xEventMessage.pvData = ( void * ) uxOption;
@@ -811,7 +811,7 @@ uint16_t usGenerateChecksum( uint16_t usSum,
     xUnionPtr xSource;
     xUnionPtr xLastSource;
     uintptr_t uxAlignBits;
-    uint32_t ulCarry = 0UL;
+    uint32_t ulCarry = 0U;
     uint16_t usTemp;
     size_t uxDataLengthBytes = uxByteCount;
 
@@ -822,7 +822,7 @@ uint16_t usGenerateChecksum( uint16_t usSum,
     /* Swap the input (little endian platform only). */
     usTemp = FreeRTOS_ntohs( usSum );
     xSum.u32 = ( uint32_t ) usTemp;
-    xTerm.u32 = 0UL;
+    xTerm.u32 = 0U;
 
     xSource.u8ptr = ipPOINTER_CAST( uint8_t *, pucNextData );
     uxAlignBits = ( ( ( uintptr_t ) pucNextData ) & 0x03U );
